requirements.txt
Flask
psycopg2-binary
pandas
scikit-learn
tensorflow
joblib
python-dotenv
gunicorn
sqlalchemy
psycopg2-binary